requirements/dev.txt
Fabric3==1.10.2.post3
pycrypto==2.6.1
